--- front/src/Test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function Test() {
-    return (
-        <p>Ceci est la page de test !</p>
-    );
-}
-
-export default Test;
