--- front/src/tailwind.config.js
@@ -1,11 +1,0 @@
-const defaultTheme = require('tailwindcss/defaultTheme')
-
-module.exports = {
-    theme: {
-      extend: {
-        fontFamily: {
-          sans: ['Inter var', ...defaultTheme.fontFamily.sans],
-        },
-      },
-    },
-  }
